Refactor GetGenesisApiData
